Refactor: adjusting location of PHASE TYPE validations

+++ src/main/java/com/electricitybill/service/impl/TipoFaseServiceImpl.java
@@ -3,10 +3,12 @@ package com.electricitybill.service.impl;
 import com.electricitybill.entity.TipoFaseEntity;
 import com.electricitybill.repository.TipoFaseRepository;
 import com.electricitybill.service.TipoFaseService;
+import jakarta.persistence.NoResultException;
 import lombok.AllArgsConstructor;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Objects;
 
 @AllArgsConstructor
 public class TipoFaseServiceImpl implements TipoFaseService {
@@ -15,11 +17,17 @@ public class TipoFaseServiceImpl implements TipoFaseService {
 
     @Override
     public List<TipoFaseEntity> findAll() {
+        if (tipoFaseRepository.findAll().isEmpty()) {
+            throw new NoResultException("Search not found!");
+        }
         return tipoFaseRepository.findAll();
     }
 
     @Override
     public TipoFaseEntity findById(int id) {
+        if (Objects.isNull(tipoFaseRepository.findById(id))) {
+            throw new NoResultException("Phase type not found!");
+        }
         return tipoFaseRepository.findById(id);
     }
 
@@ -30,11 +38,17 @@ public class TipoFaseServiceImpl implements TipoFaseService {
 
     @Override
     public boolean update(int id, TipoFaseEntity entity) throws SQLException {
+        if (Objects.isNull(tipoFaseRepository.findById(id))) {
+            throw new NoResultException("Phase type not found!");
+        }
         return tipoFaseRepository.update(id, entity);
     }
 
     @Override
     public boolean delete(int id) throws SQLException {
+        if (Objects.isNull(tipoFaseRepository.findById(id))) {
+            throw new NoResultException("Phase type not found!");
+        }
         return tipoFaseRepository.delete(id);
     }
 }
